requirements.txt
streamlit
scapy
requests
pandas
python-dotenv